chore(@clayui/modal): update Modal component stories
